30.01.2025

+++ src/app/srvoffer/actionsOffer.ts
@@ -11,24 +11,44 @@ const url = './api/offers';
   }
 
 //--------------------------------------------------------------
-async function updateRequest(uslfrmidn: number, data: SprUslFrm) {
-  await fetch(`${url}/${uslfrmidn}`, {
-    method: 'DELETE',
-    headers: {
-      'Content-Type': 'application/json',
+async function updateRequest(uslidn: number, data: SprUslFrm) {
+  const DelFlg = data.uslfrmflg;
+  const DelIdn = data.uslfrmidn;
+  //console.log("updateRequest_uslidn=",uslidn, DelFlg, DelIdn,data);
+
+  if (DelIdn) {
+    await fetch(`${url}/${DelIdn}`, {
+      method: "DELETE",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    });
+  }
+
+  //console.log("updateRequest_PrzDel=",DelFlg);
+
+  if (DelFlg) {
+  //  console.log("updateRequest_data.uslfrmflg=true=",DelIdn, data);
+    const dataPrs = {
+        uslfrmtrf: data.uslfrmtrf,
+        uslfrmhsp: data.uslfrmhsp,
+        uslminlet: Number(data.uslminlet),
+        uslmaxlet: Number(data.uslmaxlet),
+      };
+
+   //   console.log("updateRequest_POST=",url);
+
+      const response = await fetch(url, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(dataPrs),
+      });
+      return response.json();
     }
-  });
-
-  const response = await fetch(url, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(data),
-  });
-  return response.json();
-}
-
+  }
+  
 //--------------------------------------------------------------
  
  export default function useOffers() {
